pyproject.toml
[tool.poetry]
name = "task-observer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
sqlalchemy = "^2.0.38"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
itsdangerous = "^2.2.0"
jwt = "^1.3.1"
werkzeug = "^3.1.3"
python-multipart = "^0.0.20"
pytz = "^2025.1"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
